Long to int to comply with sql datatype

+++ src/main/java/org/vfl/vintago/dto/DeliveryTruckDTO.java
@@ -1,15 +1,15 @@
 package org.vfl.vintago.dto;
 
 public class DeliveryTruckDTO {
-    private final Long id;
+    private final int id;
     private final String name;
 
-    public DeliveryTruckDTO(Long id, String name) {
+    public DeliveryTruckDTO(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
