+++ src/main/java/org/vfl/vintago/dto/RouteDTO.java
@@ -4,19 +4,19 @@ import java.time.LocalDate;
 import java.util.List;
 
 public class RouteDTO {
-    private final Long id;
+    private final int id;
     private final LocalDate deliveryDate;
     private final String deliveryTruckName;
     private final List<RouteAddressDTO> addresses;
 
-    public RouteDTO(Long id, LocalDate deliveryDate, String deliveryTruckName, List<RouteAddressDTO> addresses) {
+    public RouteDTO(int id, LocalDate deliveryDate, String deliveryTruckName, List<RouteAddressDTO> addresses) {
         this.id = id;
         this.deliveryDate = deliveryDate;
         this.deliveryTruckName = deliveryTruckName;
         this.addresses = addresses;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
     public LocalDate getDeliveryDate() {
